refactor (ONB-204): update header tabs to replace Keyword Analyzer with Keyword Density; enhance index.vue with new button and animation components.
