+++ src/constants/HeaderTabs.js
@@ -2,17 +2,14 @@ const tabs = [
   {
     label: 'Home',
     route: '/',
-    icon: null
   },
   {
     label: 'Keyword Generator',
     route: '/keyword-generator',
-    icon: null
   },
   {
-    label: 'Keyword Analyzer',
-    route: '/keyword-analyzer',
-    icon: null
+    label: 'Keyword Density',
+    route: '/keyword-density',
   }
 ];
