+++ src/router/index.js
@@ -14,9 +14,9 @@ const router = createRouter({
       component: () => import('../pages/keyword-generator/KeywordGenerator.vue')
     },
     {
-      path: '/keyword-analyzer',
-      name: 'KeywordAnalyzer',
-      component: () => import('../pages/keyword-analyzer/KeywordAnalyzer.vue')
+      path: '/keyword-density',
+      name: 'KeywordDensity',
+      component: () => import('../pages/keyword-density/KeywordDensity.vue')
     }
   ]
 })
